testing mongo datapoint to prediction
