+++ Munging/mongo_to_df.py
@@ -1,4 +1,5 @@
 ''' Authors: Dylan Albrecht
+             Trent Woodbury
 
     Date: December 18, 2016
 
@@ -18,68 +19,17 @@ import pandas as pd
 import numpy as np
 from unidecode import unidecode
 from pymongo import MongoClient
-from pymongo.errors import ServerSelectionTimeoutError
 
-
-def flatten(d, parent_key='', sep='_'):
-    ''' Recursive algorithm to flatten nested dictionaries, with keys
-        created by joining with 'sep'.
-        INPUT: dict, str, str
-        OUTPUT: dict
-    '''
-    items = []
-
-    for k, v in d.items():
-        new_key = parent_key + sep + k if parent_key else k
-
-        if isinstance(v, collections.MutableMapping):
-            # Extend by sub keys -- enter recursive:
-            items.extend(flatten(v, parent_key=new_key, sep=sep).items())
-        else:
-            items.append((new_key, v))
-
-    return dict(items)
+from utils import connect_mongo
+from utils import flatten
+from utils import convert_columns
 
 
 if __name__ == '__main__':
 
     ########
     # Load
-
-    MONGO_USERNAME = os.environ['MONGO_USERNAME']
-    MONGO_PASSWORD = os.environ['MONGO_PASSWORD']
-    MONGO_HOSTNAME = os.environ['MONGO_HOSTNAME']
-
-    # Check server:
-    try:
-        address = 'mongodb://'
-        address += MONGO_USERNAME + ':'
-        address += MONGO_PASSWORD + '@'
-        address += MONGO_HOSTNAME
-        cli = MongoClient(address, serverSelectionTimeoutMS=100)
-        cli.server_info()
-
-    except ServerSelectionTimeoutError as e:
-        print "Server error!  (Is it plugged in?): "
-        print e
-        raise e
-    
-    raw = 'craft_beers_raw'
-    clean = 'craft_beers'
-
-    db = cli['beer_db']
-    cols = db.collection_names()
-
-    # Check for collections:
-    if raw not in cols:
-        print raw + ' does not exists yet! Run web-scraper first!'
-        sys.exit()
-
-    if clean not in cols:
-        print clean + ' does not exists! Run raw_to_clean.py first!'
-        sys.exit()
-
-    beer_co_clean = db['craft_beers']
+    beer_co_clean = connect_mongo()
 
     # Add entries to Pandas dataframe:
     mydata = []
@@ -94,17 +44,7 @@ if __name__ == '__main__':
 
     # Get rid of mongo id
     del df['_id']
-
-    # Take care of columns like: NaN(float) ... RwZ9MZ(unicode) ...
-    # Store as graphlab compatible: None(NoneType) ... 'RwZ9MZ'(string) ... 
-    for col in df:
-        if df[col].dtype == np.object:
-            df[col] = df[col].apply(lambda x: None if type(x) == float \
-                                                else unidecode(x))
-
-        # And convert 'float' columns:
-        df[col] = pd.to_numeric(df[col], errors='ignore')
-
+    df = convert_columns(df)
     df.drop_duplicates(inplace=True)
 
     ########
@@ -114,6 +54,4 @@ if __name__ == '__main__':
 
     with open(savefile, 'wb') as f:
         pickle.dump(df, f, protocol=pickle.HIGHEST_PROTOCOL)
-
-
